feat: add video iframe component and modals for editing history and merging metadata
- Implemented VideoIframe component for embedding videos from YouTube and Dailymotion.
- Created EditHistoryItemModal for editing user history records with relevant fields.
- Added MergeMetadataModal for merging media metadata with a specified destination ID.
- Introduced DisplayShowSeasonEpisodesModal for displaying episodes of a selected season.
- Refactored media item types and utility functions for better organization and reusability.
