+++ apps/frontend/app/components/media-item/constants.ts
@@ -1,0 +1,12 @@
+import { MediaLot } from "@ryot/generated/graphql/backend/graphql";
+
+export const JUST_WATCH_URL = "https://www.justwatch.com";
+
+export const METADATA_LOTS_WITH_GRANULAR_UPDATES = [
+	MediaLot.Show,
+	MediaLot.Anime,
+	MediaLot.Manga,
+	MediaLot.Podcast,
+];
+
+export const SECONDS_IN_MONTH = 2629800;
