+++ apps/frontend/app/components/media-item/modals/merge-metadata-modal.tsx
@@ -1,0 +1,37 @@
+import { Button, Modal, Stack, Text, TextInput, Title } from "@mantine/core";
+import { Form } from "react-router";
+import { withQuery } from "ufo";
+
+export const MergeMetadataModal = (props: {
+	opened: boolean;
+	metadataId: string;
+	onClose: () => void;
+}) => {
+	return (
+		<Modal
+			opened={props.opened}
+			onClose={props.onClose}
+			withCloseButton={false}
+			centered
+		>
+			<Form
+				replace
+				method="POST"
+				action={withQuery(".", { intent: "mergeMetadata" })}
+			>
+				<input hidden name="mergeFrom" defaultValue={props.metadataId} />
+				<Stack>
+					<Title order={3}>Merge media</Title>
+					<Text>
+						This will move all your history, reviews, and collections from the
+						source media to the destination media. This action is irreversible.
+					</Text>
+					<TextInput label="Destination media ID" name="mergeInto" required />
+					<Button type="submit" onClick={props.onClose}>
+						Submit
+					</Button>
+				</Stack>
+			</Form>
+		</Modal>
+	);
+};
